extract useSkipToContent()
